1 - add view model for KotlinBasicFragment.kt 2 - add FullExpFragment.kt
