+++ app/src/main/java/kan/kis/learnAndroidApp/Presentation/CardLearnActivity.kt
@@ -5,8 +5,10 @@ import android.content.Intent
 import androidx.appcompat.app.AppCompatActivity
 import android.os.Bundle
 import androidx.fragment.app.Fragment
-import kan.kis.learnAndroidApp.Presentation.KotlinBasic.Companion.EXTRA_KEY_FIRST_CART
-import kan.kis.learnAndroidApp.Presentation.ThreadsCardFragment.Companion.EXTRA_KEY_2_CART
+import kan.kis.learnAndroidApp.Presentation.frgments.KotlinBasicFragment
+import kan.kis.learnAndroidApp.Presentation.frgments.KotlinBasicFragment.Companion.EXTRA_KEY_FIRST_CART
+import kan.kis.learnAndroidApp.Presentation.frgments.ThreadsCardFragment
+import kan.kis.learnAndroidApp.Presentation.frgments.ThreadsCardFragment.Companion.EXTRA_KEY_2_CART
 import kan.kis.learnAndroidApp.R
 
 class CardLearnActivity : AppCompatActivity() {
@@ -26,7 +28,7 @@ class CardLearnActivity : AppCompatActivity() {
         val thisIntent = intent.getStringExtra(KEY_OPEN)
         when(thisIntent) {
             EXTRA_KEY_FIRST_CART -> {
-                openFragment(KotlinBasic())
+                openFragment(KotlinBasicFragment())
             }
 
             EXTRA_KEY_2_CART -> {
