+++ app/src/main/java/kan/kis/learnAndroidApp/Presentation/frgments/FullExpFragment.kt
@@ -1,0 +1,27 @@
+package kan.kis.learnAndroidApp.Presentation.frgments
+
+import android.os.Bundle
+import android.view.LayoutInflater
+import android.view.View
+import android.view.ViewGroup
+import android.widget.Toast
+import androidx.fragment.app.Fragment
+import kan.kis.learnAndroidApp.R
+
+class FullExpFragment: Fragment() {
+
+    override fun onCreateView(
+        inflater: LayoutInflater,
+        container: ViewGroup?,
+        savedInstanceState: Bundle?
+    ): View? {
+        val view = inflater.inflate(R.layout.full_exp_fragment, container, false)
+        return view
+    }
+
+
+    override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
+        Toast.makeText(requireContext(), "This is work", Toast.LENGTH_SHORT).show()
+    }
+
+}
