+++ app/src/main/java/kan/kis/learnAndroidApp/Presentation/viewModels/KotlinBasicViewModel.kt
@@ -1,45 +1,29 @@
-package kan.kis.learnAndroidApp.Presentation
+package kan.kis.learnAndroidApp.Presentation.viewModels
 
-import android.os.Bundle
 import android.util.Log
-import androidx.appcompat.app.AppCompatActivity
+import androidx.lifecycle.LiveData
+import androidx.lifecycle.MutableLiveData
+import androidx.lifecycle.ViewModel
 import com.google.firebase.database.DataSnapshot
 import com.google.firebase.database.DatabaseError
 import com.google.firebase.database.DatabaseReference
 import com.google.firebase.database.ValueEventListener
 import com.google.firebase.database.ktx.database
 import com.google.firebase.ktx.Firebase
-import kan.kis.learnAndroidApp.Presentation.Adapters.CardAdapter
-import kan.kis.learnAndroidApp.Presentation.frgments.KotlinBasicFragment.Companion.EXTRA_KEY_FIRST_CART
-import kan.kis.learnAndroidApp.Presentation.frgments.ThreadsCardFragment.Companion.EXTRA_KEY_2_CART
-import kan.kis.learnAndroidApp.databinding.ActivityMainBinding
+import kan.kis.learnAndroidApp.Presentation.Pojo.CardItem
+import kan.kis.learnAndroidApp.Presentation.Pojo.TypeItem
 
-class MenuActivity : AppCompatActivity() {
+class KotlinBasicViewModel: ViewModel() {
 
+    private val TAG = "KotlinBasicViewModelTAG"
 
-
-    lateinit var binding: ActivityMainBinding
-
-    lateinit var adapterCard: CardAdapter
-
-    private val TAG = "MenuActivityTAG"
+    private val _firebaseDataBase = MutableLiveData<MutableList<CardItem>>()
+    val firebaseDataBase: LiveData<MutableList<CardItem>>
+        get() = _firebaseDataBase
 
     private lateinit var database: DatabaseReference
 
-
-    override fun onCreate(savedInstanceState: Bundle?) {
-        super.onCreate(savedInstanceState)
-        binding = ActivityMainBinding.inflate(layoutInflater)
-        setContentView(binding.root)
-
-        menuListeners()
-
-        checkFirebaseDataBase()
-
-    }
-
-    // get info from firebase
-    private fun checkFirebaseDataBase() {
+    fun checkFirebaseDataBase(mutableList: MutableList<CardItem>): MutableList<CardItem> {
         database = Firebase.database.reference
         // just check all child in fire base
         val myRef = database.child("key").child("kotlinBasic").get().addOnSuccessListener {
@@ -57,6 +41,15 @@ class MenuActivity : AppCompatActivity() {
                     child.children.forEach {
                         Log.d(TAG, "ChildrenChild: ${it.getValue()}")
                         Log.d(TAG, "City state: ${it.child("title")}")
+                        mutableList.add(
+                            CardItem(
+                                it.child("id").value.toString().toInt(),
+                                TypeItem.HEADER,
+                                it.child("body").value.toString(),
+                                it.child("title").value.toString()
+                            )
+                        )
+                        _firebaseDataBase.value = mutableList
                     }
                 }
             }
@@ -68,26 +61,11 @@ class MenuActivity : AppCompatActivity() {
 
         database.addValueEventListener(valueEventListener)
 
-
-//        myRef.setValue("hello lol")
-//
         myRef.addOnCompleteListener {
             Log.d(TAG, "Complete listeners: $it")
         }
+
+
+        return mutableList
     }
-
-    private fun menuListeners() {
-        binding.mmItem1.setOnClickListener {
-            startActivity(CardLearnActivity().newIntent(this, EXTRA_KEY_FIRST_CART))
-        }
-
-        binding.mmItem2.setOnClickListener {
-            startActivity(CardLearnActivity().newIntent(this, EXTRA_KEY_2_CART))
-        }
-
-
-
-//        binding.mmItem4.setOnClickListener().
-    }
-
 }
