+++ app/src/main/res/layout/activity_card_learn.xml
@@ -2,7 +2,7 @@
 <androidx.fragment.app.FragmentContainerView xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:app="http://schemas.android.com/apk/res-auto"
     xmlns:tools="http://schemas.android.com/tools"
-    android:name="kan.kis.learnAndroidApp.Presentation.KotlinBasic"
+    android:name="kan.kis.learnAndroidApp.Presentation.frgments.KotlinBasicFragment"
     android:id="@+id/fragment_card_container"
     tools:layout="@layout/cart_fragment"
     android:layout_width="match_parent"
